+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on Keep a Changelog,
 and this project adheres to Semantic Versioning.
 
+## [0.0.10] - 2026-01-29
+
+### Changed
+
+- `user/delete` updated response
+
+### Added
+
+- Added users interests delete and add route
+
 ## [0.0.9] - 2026-01-29
 
 ### Changed